backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
pydantic==2.9.2
pydantic-settings==2.5.2
python-multipart==0.0.12
sqlmodel==0.0.22
httpx==0.27.2
pypdf==5.0.1
pdfplumber==0.11.4
python-dotenv==1.0.1
google-genai==0.3.0
google-auth==2.35.0
google-cloud-storage==2.18.2
tenacity==9.0.0
structlog==24.4.0
pytest==8.3.3
pytest-asyncio==0.24.0
